+++ packages/react/bunchee.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'bunchee'
-
-export default defineConfig({
-  entries: ['src/index.ts'],
-  external: ['react', 'react-dom', 'react/jsx-runtime'],
-  dts: true,
-})